Struct wasmer_engine_dylib::engine::DylibEngine

pub struct DylibEngine {
    inner: Arc<Mutex<DylibEngineInner>>,
    target: Arc<Target>,
    engine_id: EngineId,
}

A WebAssembly Dylib Engine.

Fields
inner: Arc<Mutex<DylibEngineInner>>target: Arc<Target>

The target for the compiler

engine_id: EngineId
Implementations

impl DylibEngine

pub fn new(
compiler: Box<dyn Compiler>,
target: Target,
features: Features
) -> Self

Create a new DylibEngine with the given config

pub fn headless() -> Self

Create a headless DylibEngine

A headless engine is an engine without any compiler attached. This is useful for assuring a minimal runtime for running WebAssembly modules.

For example, for running in IoT devices where compilers are very expensive, or also to optimize startup speed.

Important

Headless engines can't compile or validate any modules, they just take already processed Modules (via Module::serialize).

pub fn set_deterministic_prefixer<F>(&mut self, prefixer: F) where
F: Fn(&[u8]) -> String + Send + 'static,

Sets a prefixer for the wasm module, so we can avoid any collisions in the exported function names on the generated shared object.

This, allows us to rather than have functions named wasmer_function_1 to be named wasmer_function_PREFIX_1.

Important

This prefixer function should be deterministic, so the compilation remains deterministic.

pub(crate) fn inner(&self) -> MutexGuard<'_, DylibEngineInner>

pub(crate) fn inner_mut(&self) -> MutexGuard<'_, DylibEngineInner>