Struct wasmer_engine_dylib::engine::DylibEngineInner

pub struct DylibEngineInner {
    compiler: Option<Box<dyn Compiler>>,
    features: Features,
    signatures: SignatureRegistry,
    func_data: Arc<FuncDataRegistry>,
    prefixer: Option<Box<dyn Fn(&[u8]) -> String + Send>>,
    is_cross_compiling: bool,
    linker: Linker,
    libraries: Vec<Library>,
}

The inner contents of DylibEngine

Fields
compiler: Option<Box<dyn Compiler>>

The compiler

features: Features

The WebAssembly features to use

signatures: SignatureRegistry

The signature registry is used mainly to operate with trampolines performantly.

func_data: Arc<FuncDataRegistry>

The backing storage of VMFuncRefs. This centralized store ensures that 2 functions with the same VMCallerCheckedAnyfunc will have the same VMFuncRef. It also guarantees that the VMFuncRefs stay valid until the engine is dropped.

prefixer: Option<Box<dyn Fn(&[u8]) -> String + Send>>

The prefixer returns the a String to prefix each of the functions in the shared object generated by the DylibEngine, so we can assure no collisions.

is_cross_compiling: bool

Whether the Dylib engine will cross-compile.

linker: Linker

The linker to use.

libraries: Vec<Library>

List of libraries loaded by this engine.

Implementations

impl DylibEngineInner

pub fn compiler(&self) -> Result<&dyn Compiler, CompileError>

Gets the compiler associated to this engine.

pub(crate) fn get_prefix(&self, bytes: &[u8]) -> String

pub(crate) fn features(&self) -> &Features

pub fn validate<'data>(&self, data: &'data [u8]) -> Result<(), CompileError>

Validate the module

pub fn signatures(&self) -> &SignatureRegistry

Shared signature registry.

pub(crate) fn func_data(&self) -> &Arc<FuncDataRegistry>

Shared func metadata registry.

pub(crate) fn is_cross_compiling(&self) -> bool

pub(crate) fn linker(&self) -> Linker

pub(crate) fn add_library(&mut self, library: Library)