Module wasmer_engine_dylib::trampoline

Trampolines for libcalls.

This is needed because the target of libcall relocations are not reachable through normal branch instructions.

There is an additional complexity for dynamic libraries: we can’t just import the symbol from the host executable because executables don’t export dynamic symbols (it’s possible but requires special linker options).

Instead, we export a table of function pointers in the data section which is manually filled in by the runtime after the dylib is loaded.

Constants

AARCH64_TRAMPOLINE

WASMER_TRAMPOLINES_SYMBOL

Symbol exported from the dynamic library which points to the trampoline table.

X86_64_TRAMPOLINE

Functions

emit_trampoline

emit_trampolines

Emits the libcall trampolines and table to the object file.

fill_trampoline_table⚠

Fills in the libcall trampoline table at the given address.