Crate wasmer_engine_dylib

Dylib Engine for Wasmer compilers.

Given a compiler (such as CraneliftCompiler or LLVMCompiler) it generates a dylib/shared object file (.so or .dylib depending on the target), saves it temporarily to disk and uses it natively via dlopen and dlsym (using the libloading library).

Modules

artifact
Define DylibArtifact to allow compiling and instantiating to be done as separate steps.

builder

engine
Dylib Engine.

serialize

Structs

Dylib
The Dylib builder

DylibArtifact
A compiled Wasm module, ready to be instantiated.

DylibEngine
A WebAssembly Dylib Engine.

Constants

VERSION
Version number of this crate.